+++ client/src/components/Footer/addressesSyG.json
@@ -1,0 +1,18 @@
+
+[
+    {
+        "nombre": "av. Granaderos 2205, Local 2. Esquina Felix Hoyos, al frente de PDI",
+        "latitud": "-22.458819243014876",
+        "longitud": "-68.92856193508142"
+    },
+    {
+        "nombre": "Av. Arturo Prat 1597-B Frente al Hospital Carlos Cisternas",
+        "latitud": "-22.446121700013478",
+        "longitud": "-68.91924964324909"
+    },
+    {
+        "nombre": "Av. Grecia 2510, 1392007 Calama, Antofagasta",
+        "latitud": "-22.455721547785448",
+        "longitud": "-68.93388475804485"
+    }
+]
