+++ client/src/components/Footer/linkStyles.jsx
@@ -1,0 +1,16 @@
+
+const linkStyles = {
+    color: 'var(--font-link-color)',
+    fontFamily: 'var(--font-body)',
+    fontSize: '20px',
+    '&.active': {
+        textDecoration: 'none',
+        color: 'var(--font-footer-color3)',
+    },
+    "&:hover": {
+        color: "var(--font-footer-color2)",
+    },
+};
+
+
+export { linkStyles};
