+++ client/src/components/Footer/linkStyles2.jsx
@@ -1,0 +1,14 @@
+const linkStyles2 = {
+    color: 'var(--font-link-color)',
+    fontFamily: 'var(--font-body)',
+    fontSize: '20px',
+    '&.active': {
+        textDecoration: 'none',
+        color: 'var(--font-footer-color3)',
+    },
+    "&:hover": {
+        color: "var(--font-link-color)",
+    },
+};
+
+export { linkStyles2};
